Move shortest path and PageRank to bagel.examples
